+++ internal/utils.go
@@ -1,7 +1,6 @@
 package internal
 
 import (
-	"encoding/base64"
 	"errors"
 	"fmt"
 	"os"
@@ -19,9 +18,15 @@ func IsEmpty(val string) bool {
 	return len(strings.TrimSpace(val)) == 0
 }
 
-// FormatTime formats the time using `02-01-2006T15:04:05.999999999` layout.
-func FormatTime(t time.Time) string {
-	return t.Format("02-01-2006T15:04:05.999999999")
+// FormatTimeForHuman formats the time using `02 Jan 2006 15:04:05.999999999` layout.
+func FormatTimeForHuman(t time.Time) string {
+	return t.Format("02 Jan 2006 15:04:05.999999999")
+}
+
+// FormatTimeForMachine formats the time using `2006-01-02T15:04:05.999999999` layout.
+func FormatTimeForMachine(t time.Time) string {
+	// yyyy-mm-dd
+	return t.Format("2006-01-02T15:04:05.999999999")
 }
 
 // NotFoundError constructs a new NotFound error for the specified entity.
@@ -31,29 +36,6 @@ func NotFoundError(entity, filterName string, ex *regexp.Regexp) error {
 		msg += fmt.Sprintf(" You might need to tweak the %s filter (%s).", filterName, ex.String())
 	}
 	return errors.New(msg)
-}
-
-// FormatTimeUTC formats the time using `02-01-2006T15:04:05.999999999` layout and adds the "UTC" suffix to the end.
-func FormatTimeUTC(t time.Time) string {
-	return FormatTime(t) + " UTC"
-}
-
-// PrependTimestamp prefixes the input with a timestamp string in `02-01-2006T15:04:05.999999999 UTC` format.
-func PrependTimestamp(ts time.Time, in []byte) []byte {
-	return append([]byte(fmt.Sprintf("%s\n", FormatTimeUTC(ts))), in...)
-}
-
-// PrependTopic prefixes the input with the topic name.
-func PrependTopic(topic string, in []byte) []byte {
-	return append([]byte(fmt.Sprintf("%s\n", topic)), in...)
-}
-
-// PrependKey prefixes the input with the partition key in `PN: Hex` format where N is the partition number.
-func PrependKey(key []byte, partition int32, in []byte, b64 bool) []byte {
-	if b64 {
-		return append([]byte(fmt.Sprintf("P%d: %s\n", partition, base64.StdEncoding.EncodeToString(key))), in...)
-	}
-	return append([]byte(fmt.Sprintf("P%d: %X\n", partition, key)), in...)
 }
 
 // WaitForCancellationSignal waits for the user to press Ctrl+C.